Updated fings

+++ src/main/CharacterInfo.java
@@ -4,7 +4,7 @@ public class CharacterInfo {
 
   public int character_level;
   public CharacterClass character_class;
-  public int max_hp;
+  public double max_hp;
   public int max_ap;
   public int exp;
   public int attack;
